Diabetes Risk

Subset of survey data collected by the US National Center for Health Statistics (NCHS). The original data was based on home interviews of about 5,000 people per years, from 1999-2004.

Format

A data frame with 9096 observations on the following 23 variables.

sex

"male" or "female"

age

age of subject in years

pregnant

"yes" or "no"

ethnicity

Mexican American, Other Hispanic, Non-Hispanic White, Non-Hispanic Black, or Other/Multi

smoker

"yes" or "no"

diabetic

"yes" or "no"

height

height (meters)

weight

weight (kilograms)

waist

waist circumference (meters)

wci

the proposed body shape index

bmi

body mass index

ptfat

percent trunk fat

tfat

mass of trunk fat

lfat

limb fat

llean

limb lean tissue

lbmi

lean-tissue only BMI

fbmi

fat-only BMI

bbmi

bone BMI

pfat

percent fat

bmd

bone mineral density

fmhm_other

Framingham risk score

hdl

HDL cholesterol

chol

cholesterol (LDL?)

bps

systolic blood pressure, mmHg

bpd

diastolic blood pressure, mmHg

income

ratio of family income to poverty threshold. (5 stands for a ratio greater than or equal to 5)

Source

Modified from NCHS in package DataComputing. The original data is from NHANES, the National Health and Nutrition Survey. See http://wwwn.cdc.gov/nchs/nhanes/search/nhanes03_04.aspx# for more infromation.