Homonegativity in Greek Life

Georgetown College students surveyed their peers on attitudes about LBGT issues and persons.

Format

A data frame with 75 observations on the following 22 variables.

Gender

LGBTPlus

Age

Race

School.System

LGBTPlusFamily

LGBTPlusFamily.Proximity

LGBTPlusFriends

College.Year

Sports

Greek

Q1G

Q2G

Q3G

Q4E

Q5V

Q6R

Q7V

Q8V

Q9E

Q10V

Q11E

Source

Hunter Gatewood, Molly Dixon, Almond Bailey. Class: PSY 311, Fall 2014. Instructor: Dr. Regan Lookadoo. For survey form see ../doc/lbgt_likert.pdf.

Note

From Hunter Gatewood (email): "It was coded mainly based upon the numbers stated within the questions. For question 7 (on the front), if they answered no for question 6 (front), we coded it as 5. For question 11, we had to separate it between 'for a short time' (1-3 semesters) and 'for a longer time (4+ semesters) with coding (1,2)."