Can You Eat This Mushroom?

Subset of data from a study on edibility of mushrroms. The individual mushrooms come from 23 species of gilled mushrooms in the Agaricus and Lepiota Family. The aim is to come up with a rule for predicting, on the basis of an individual mushroom's characteristics, whether or not the mushroom is edible. Remaining data is held back for evaluation of proposed rules.

Format

A data frame with 5891 observations on the following 23 variables.

class

Whether the mushroom is edible or poisonous.

cap.shape

cap.surface

cap.color

bruises

Whether or not the mushroom is bruised.

odor

gill.attachment

gill.spacing

gill.size

gill.color

stalk.shape

stalk.root

stalk.surface.above.ring

stalk.surface.below.ring

stalk.color.above.ring

stalk.color.below.ring

veil.type

veil.color

ring.number

ring.type

spore.print.color

population

habitat

Source

A sample from of mushroom records drawn from The Audubon Society Field Guide to North American Mushrooms (1981). G. H. Lincoff (Pres.), New York: Alfred A. Knopf, Original data contributed by Jeffrey Schlimmer to the UCI Machine Learning Repository (http://archive.ics.uci.edu/ml, Irvine, CA: University of California, School of Information and Computer Science. See http://archive.ics.uci.edu/ml/datasets/Mushroom.